+++ app/api/news/[newsId]/route.js
@@ -2,6 +2,9 @@ import moveUploadFile from "@/helper/moveUploadFile";
 import prisma from "@/helper/prismaInit";
 import { NextResponse } from "next/server";
 import cloudinary from "@/helper/cloudinary";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/lib/auth";
+import { getToken } from "next-auth/jwt";
 
 export const dynamic = "force-dynamic"; // defaults to force-static
 
@@ -28,6 +31,17 @@ export async function GET(request, { params }) {
 }
 
 export async function PUT(request, { params }) {
+  const token = await getToken({ req: request });
+
+  if (!token) {
+    return NextResponse.json(
+      {
+        success: false,
+        message: "You are not logged in",
+      },
+      { status: 401 }
+    );
+  }
   try {
     const formData = await request.formData();
     const file = formData.get("file");
@@ -68,6 +82,17 @@ export async function PUT(request, { params }) {
 }
 
 export async function DELETE(request, { params }) {
+  const token = await getToken({ req: request });
+
+  if (!token) {
+    return NextResponse.json(
+      {
+        success: false,
+        message: "You are not logged in",
+      },
+      { status: 401 }
+    );
+  }
   try {
     const post = await prisma.post.delete({
       where: {
